(AuthedSection) split out authedSection into own component
