+++ scripts/app.js
@@ -4,9 +4,6 @@ var Router = ReactRouter.Router;
 var Route = ReactRouter.Route;
 var Link = ReactRouter.Link;
 var history = require('react-router/lib/BrowserHistory').history;
-var Sidebar = require('react-sidebar');
-var Header = require('./components/Header.jsx');
-var NavigationList = require('./components/Navigation.jsx');
 var injectTapEventPlugin = require('react-tap-event-plugin');
 var mui = require('material-ui');
 let ThemeManager = new mui.Styles.ThemeManager();
@@ -18,60 +15,9 @@ var InspectionReport = require('./components/InspectionReport.jsx');
 var RtaForm = require('./components/RtaForm.jsx');
 var Messages = require('./components/Messages.jsx');
 var Login = require('./components/Login.jsx');
+var AuthedSection = require('./components/AuthedSection.jsx');
 
 require('../styles/main.scss');
-
-var AuthedSection = React.createClass({
-  propTypes: {
-    children: React.PropTypes.any,
-  },
-
-  getInitialState() {
-    return {
-      isSidebarOpen: false,
-      isSidebarDocked: true,
-    };
-  },
-
-  componentWillMount() {
-    this.mql = window.matchMedia(`(min-width: 800px)`);
-    this.mql.addListener(this.mediaQueryChanged);
-    this.setState({ isSidebarDocked: this.mql.matches });
-  },
-
-  componentWillUnmount() {
-    this.mql.removeListener(this.mediaQueryChanged);
-  },
-
-  mediaQueryChanged() {
-    this.setState({ isSidebarDocked: this.mql.matches });
-  },
-
-  onSetSidebarOpen(isOpen) {
-    this.setState({ isSidebarOpen: isOpen });
-  },
-
-  openMenu() {
-    this.setState({ isSidebarOpen: true });
-  },
-
-  render() {
-    return (
-      <div>
-        <Sidebar sidebar={<NavigationList />}
-                 open={this.state.isSidebarOpen}
-                 docked={this.state.isSidebarDocked}
-                 onSetOpen={this.onSetSidebarOpen}>
-          <Header onMenuClick={this.openMenu}
-                  isMenuDocked={this.state.isSidebarDocked}/>
-          {this.props.children}
-        </Sidebar>
-      </div>
-    );
-  }
-});
-
-
 
 var App = React.createClass({
   componentWillMount() {
